Remove parameterless constructor constraints.
Managed objects from native pointers are now created without running a constructor. That way the preexisting native object doesn't get reinitialized.

There is now an attribute that will be used to specify the native size of a class, and will be used to allocate memory in order for the class to interop with native code.